Split the modes, schema and surface text out of the subagent body
The last three concerns leave the entry point: the four ways a request runs
with the project-agent gate they share, the tool schema and the types derived
from it, and the plain strings /tasks, /agents and the completion notice read
out. The bounded fan-out helper moves with the parallel runner that uses it.

What remains is the schema, the dispatch and the session hooks, 303 lines from
1693. The file header is now a map of where each concern went.

Pure move: no behavior change.
